requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymilvus==2.3.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
boto3==1.34.10
python-dotenv==1.0.0
gunicorn==21.2.0
